resume-analyzer/requirements.txt
flask
python-dotenv
python-docx
PyPDF2
google-generativeai
markupsafe
xhtml2pdf
flask-login
flask-sqlalchemy